Update README.md
argest N numbers

Let M be an N × N integer matrix, where each element in M is distinct, ensuring that no two elements are equal. Propose an algorithm that efficiently identifies the N largest elements among the elements of M. Furthermore, provide a proof of the correctness of this algorithm. Note that the algorithm does not need to preserve M’s initial state.


Example

Input

5

// N

1591325